fix(formatMoney): move file to utils folder and delete helpers folder
